audio: Improve tonegenerator examples
This fixes some glitches in the generated wave, mostly by lowering
the sample rate from 24000 to 8000. It also adds a function to get
the radians increment for the wave from hz, which bases off the rate.
This means it attempts to make the frequency consistent between
different sample rates but it does poorly on rates higher than 8000.